refactor: Reorganizes project
