+++ src/schemaver/diffs/__init__.py
@@ -9,11 +9,12 @@ __all__ = [
     "StringValidationDiff",
     "NumericValidationDiff",
     "ObjectValidationDiff",
+    "CoreField",
 ]
 
 from schemaver.diffs.array import ArrayValidationDiff
 from schemaver.diffs.base import BaseDiff
-from schemaver.diffs.core import CoreValidationDiff
+from schemaver.diffs.core import CoreField, CoreValidationDiff
 from schemaver.diffs.metadata import MetadataDiff
 from schemaver.diffs.numeric import NumericValidationDiff
 from schemaver.diffs.object import ObjectValidationDiff
