+++ tests/unit_tests/diffs/test_array.py
@@ -2,14 +2,11 @@
 
 import pytest
 
-from schemaver.lookup import (
-    ChangeLevel,
-    InstanceType,
-    ArrayField,
-    StringField,
-    NumericField,
-)
-from schemaver.property import Property
+from schemaver.changelog import ChangeLevel
+from schemaver.diffs.string import StringField
+from schemaver.diffs.numeric import NumericField
+from schemaver.diffs.array import ArrayField
+from schemaver.property import InstanceType, Property
 
 from tests.unit_tests.diffs.helpers import (
     assert_changes,
