+++ tests/unit_tests/diffs/test_string.py
@@ -2,13 +2,10 @@
 
 import pytest
 
-from schemaver.lookup import (
-    ChangeLevel,
-    InstanceType,
-    StringField,
-    ArrayField,
-)
-from schemaver.property import Property
+from schemaver.changelog import ChangeLevel
+from schemaver.diffs.array import ArrayField
+from schemaver.diffs.string import StringField
+from schemaver.property import InstanceType, Property
 
 from tests.unit_tests.diffs.helpers import (
     assert_changes,
